notification_service/requirements.txt
fastapi
uvicorn
motor                     # Async MongoDB driver
pydantic
python-jose               # For JWT handling
httpx
PyJWT
pika
celery[redis]